% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decimalplaces}
\alias{decimalplaces}
\title{Get number of decimal places}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{a number}
}
\value{
the number of decimal places
}
\description{
Get number of decimal places
}
