% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{discrete_gradient_pal}
\alias{discrete_gradient_pal}
\title{Get discrete gradient palette}
\usage{
discrete_gradient_pal(colours, bins = 5)
}
\arguments{
\item{colours}{colors}

\item{bins}{bin to group numeric values into}
}
\value{
discrete gradient palette
}
\description{
Get discrete gradient palette
}
