% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_congress_map}
\alias{get_congress_map}
\title{Get Congressional Map based on a congressional year}
\usage{
get_congress_map(cong = 114)
}
\arguments{
\item{cong}{the specific congress number}
}
\value{
congressional map
}
\description{
Get Congressional Map based on a congressional year
}
