% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_congress_map_by_year}
\alias{get_congress_map_by_year}
\title{Get Congressional Map by Election Year}
\usage{
get_congress_map_by_year(year)
}
\arguments{
\item{year}{election year}
}
\value{
congressional map associated with that election year
}
\description{
Get Congressional Map by Election Year
}
