% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_congress_number}
\alias{get_congress_number}
\title{Get Congress Number}
\usage{
get_congress_number(election_year)
}
\arguments{
\item{election_year}{election year}
}
\value{
congressional number for which map the election that year is based on.
}
\description{
Get Congress Number
}
