% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congressional_vote_data.R
\name{get_congressional_vote_data}
\alias{get_congressional_vote_data}
\title{Get Congressional Vote Data}
\usage{
get_congressional_vote_data(re_process = FALSE)
}
\arguments{
\item{re_process}{whether to re process the data}
}
\value{
processed congressional vote data
}
\description{
This function gets and prepares congressional votes data
from 1976 to 2020.
}
\details{
Data features returned are:
\itemize{
\item year, state, state_po (abbreviation), state_fips (FIPS),
\item district (number), totalvotes, DEMOCRAT (# of votes), REPUBLICAN (# of votes)
\item DEMOCRAT_PRCT (\%), REPUBLICAN_PRCT (\%), SPREAD_DR (DEMOCRAT_PRCT - REPUBLICAN_PRCT)
\item TOTAL_TW (total two-way), DEMOCRAT_PRCT_TW (democratic percentage two-way \%), REPUBLICAN_PRCT_TW
\item SPREAD_DR_TW (DEMOCRAT_PRCT_TW - REPUBLICAN_PRCT_TW)
}
}
