% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_filtered_congress_data}
\alias{get_filtered_congress_data}
\title{Get filtered congressional result data}
\usage{
get_filtered_congress_data(vote_df, chosen_state, chosen_year)
}
\arguments{
\item{vote_df}{data containing congressional vote results}

\item{chosen_state}{a chosen state}

\item{chosen_year}{a chosen year}
}
\value{
filtered congressional vote result
}
\description{
Get filtered congressional result data
}
