% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congressional_vote_data.R
\name{get_generic_ballot_data}
\alias{get_generic_ballot_data}
\title{Get Generic Ballot Data}
\usage{
get_generic_ballot_data(vote_df)
}
\arguments{
\item{vote_df}{vote data}
}
\value{
generic ballot data
}
\description{
Get Generic Ballot Data
}
