% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_two_party_spread_plots}
\alias{get_two_party_spread_plots}
\title{Get Plots}
\usage{
get_two_party_spread_plots(
  chosen_year,
  vote_data,
  chosen_state,
  tooltips = TRUE
)
}
\arguments{
\item{chosen_year}{a chosen year}

\item{vote_data}{data containing congressional vote results}

\item{chosen_state}{a chosen state}
}
\value{
plots
}
\description{
Get Plots
}
