% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vote_shares}
\alias{plot_vote_shares}
\title{Plot Vote Share}
\usage{
plot_vote_shares(vote_df, chosen_state)
}
\arguments{
\item{vote_df}{vote dataframe}

\item{chosen_state}{chosen state}
}
\value{
vote shares by districts
}
\description{
Plot Vote Share
}
