% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plots_two_party_spreads}
\alias{plots_two_party_spreads}
\title{Plot Two-party Spreads}
\usage{
plots_two_party_spreads(
  vote_df,
  chosen_state,
  start_year = 1976,
  end_year = 2020,
  tooltips = TRUE
)
}
\arguments{
\item{vote_df}{vote data}

\item{chosen_state}{chosen state}

\item{start_year}{election start year}

\item{end_year}{election end year}
}
\description{
Plot Two-party Spreads
}
