% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_fill_discrete_gradient}
\alias{scale_fill_discrete_gradient}
\title{Fill discrete gradient}
\usage{
scale_fill_discrete_gradient(
  ...,
  colours,
  bins = 5,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill",
  colors
)
}
\arguments{
\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\value{

}
\description{
Fill discrete gradient
}
